requirements.txt
tensorflow-cpu==2.12.0
numpy==1.22.4
flask==2.2.2
flask_cors==3.0.10
gunicorn==20.1.0
requests==2.30.0
